Move binary suffix to new units package
A `units` package makes sense here and sounds better. Code has no other
changes.
