+++ units/binary_suffix.go
@@ -1,4 +1,4 @@
-package strings
+package units
 
 import (
 	"fmt"
