+++ units/binary_suffix_test.go
@@ -1,4 +1,4 @@
-package strings
+package units
 
 import (
 	"github.com/stretchr/testify/assert"
